syntax and other corrections * apache2 installation command isolated from other install commands * removed atom from list of software. deal with it later * before continuing with apache2 config, test for /var/www   * if it doesn't exist, exit the script with warning * only configure a www folder base for axemoor if one doesn't exist yet * only create a .conf file for axemoor in sites-available if one doesn't exist * change the helper text for MariaDB/MySQL setup to display in yellow
non-fatal errors that aren't fixed yet:
* checking for existence of git repos before attempting to clone

+++ raspbian-init.sh
@@ -6,42 +6,60 @@
 # Raspbian Edition
 
 # First, some housekeeping
-echo "Beginning system update and software install"
-sudo apt update
-sudo apt upgrade -y
+echo "Beginning software update"
+sudo apt update > /dev/null
+sudo apt upgrade > -y /dev/null
 echo "Software update complete"
-$install1="apache2 git systemctl atom"
-$install2="php libcache2-mod-php"
+
+$install2="php libcache2-mod-php git systemctl"
 $install3="mariadb-server"
-sudo apt install "$install1" -y
+sudo apt install apache2 -y > /dev/null
+echo "Done with Apache2"
+echo "Enable & Start SSH"
 sudo systemctl enable ssh
 sudo systemctl start ssh
-echo "Done installing Apache2"
+echo "Done with SSH"
 # Apache Config
 echo "Apache config starting now"
-sudo mkdir /var/www/axemoor/public_html -p
-sudo chown -R $USER:$USER /var/www/axemoor/public_html
-sudo chmod -R 755 /var/www
-sudo touch /etc/apache2/sites-available/axemoor.net.conf
-$newconf="/etc/apache2/sites-available/axemoor.net.conf"
-echo '<VirtualHost *:80>' >> $newconf
-echo '    ServerAdmin webminister@axemoor.net' >> $newconf
-echo '    ServerName axemoor.net' >> $newconf
-echo '    ServerAlias www.axemoor.net' >> $newconf
-echo '    DocumentRoot /var/www/html' >> $newconf
-echo '    ErrorLog ${APACHE_LOG_DIR}/error.log' >> $newconf
-echo '    CustomLog ${APACHE_LOG_DIR}/access.log combined' >> $newconf
-echo '</VirtualHost>' >> $newconf
-sudo a2ensite axemoor.net.conf
-sudo a2dissite 000-default.conf
+# Check for /var/www. If not found, apache2 install failed 
+# and we need to abort the install.
+if [ ! -d "/var/www" ];
+  then echo "Apache didn't install correctly. Aborting." && exit
+fi
+# Check for /var/www/axemoor. If not exist, make
+if [ ! -d "/var/www/axemoor" ];
+  then
+  sudo mkdir /var/www/axemoor/public_html -p
+  sudo chown -R $USER:$USER /var/www/axemoor/public_html
+  sudo chmod -R 755 /var/www
+fi
+# Check for axemoor.net.conf in sites-available. If not exist, make.
+if [ ! -f "/etc/apache2/sites-available/axemoor.net.conf" ];
+  then
+  sudo touch /etc/apache2/sites-available/axemoor.net.conf
+  $newconf="/etc/apache2/sites-available/axemoor.net.conf"
+  echo '<VirtualHost *:80>' >> $newconf
+  echo '    ServerAdmin webminister@axemoor.net' >> $newconf
+  echo '    ServerName axemoor.net' >> $newconf
+  echo '    ServerAlias www.axemoor.net' >> $newconf
+  echo '    DocumentRoot /var/www/html' >> $newconf
+  echo '    ErrorLog ${APACHE_LOG_DIR}/error.log' >> $newconf
+  echo '    CustomLog ${APACHE_LOG_DIR}/access.log combined' >> $newconf
+  echo '</VirtualHost>' >> $newconf
+  sudo a2ensite axemoor.net.conf
+  sudo a2dissite 000-default.conf
+fi
+
 echo "Attempting to restart Apache. Hang on."
 sudo systemctl restart apache2
-#echo "Apache status is"
-#sudo systemctl status apache2
+echo "Apache status is"
+sudo systemctl status apache2
 echo "Done configuring Apache. Now installing additional software."
-sudo apt install "$install2" -y #php
+sudo apt install "$install2" -y #php libcache2-mod-php git systemctl
 sudo apt-get install "$install3" -y #mariadb-server
-echo "This next section will ask you for input."
+RED='\033[0;31m'
+NC='\033[0m'
+echo "${red}This next section will ask you for input."
 echo "Recommended answers are provided here. Copy and paste this list"
 echo "to someplace where you can reference it while you answer these"
 echo "questions."
@@ -51,7 +69,7 @@ echo "2. Change the root password: Choose 'n'"
 echo "3. Remove anonymous users: Choose 'y'"
 echo "4. Disallow root login remotely: Choose 'y'"
 echo "5. Remove test database and access: Choose 'y'"
-echo "6. Reload privilege tables: Choose 'y'"
+echo "6. Reload privilege tables: Choose 'y'${NC}"
 sudo mysql_secure_installation
 sudo apt install php-mysql -y
 sudo service apache2 restart
